Clean up the fuzzing defines
This tries to move as much of the fuzzing code to the cfs-fuzzing.h
header as possible.
