+++ kernel/cfs-reader.h
@@ -5,16 +5,6 @@
 
 #define EFSCORRUPTED EUCLEAN /* Filesystem is corrupted */
 
-#ifdef FUZZING
-static inline void *ERR_CAST(const void *ptr)
-{
-	return (void *)ptr;
-}
-#define ERR_PTR(x) ((void *)((long)x))
-#define PTR_ERR(x) ((long)x)
-#define IS_ERR(x) ((unsigned long)(void *)(x) >= (unsigned long)-4096)
-#endif
-
 struct cfs_context_s;
 
 struct cfs_context_s *cfs_create_ctx(const char *descriptor_path,
